factor out shared code
